feat(tones): changed to color
This is a breaking change. In Button and IconButton now the available values for color prop are
"primary" | "secondaryB" and "inherit" (only for IconButton). The variant prop is renamed in
ChevronButton, ChevronLink and ChevronLinkRenderer to "color". The value is also matched with the
allowed values of Button "color". In Preloader and Spinner the tone prop is renamed to "color". In
SvgIcon, the prop "tone" is deprecated and "color" prop is added instead.